backend/requirements.txt
Flask
python-dotenv
pyodbc
azure-identity
azure-keyvault-secrets
gunicorn
